+++ extension/data/init.ts
@@ -34,10 +34,10 @@ import * as TBApi from './tbapi';
 import * as TBCore from './tbcore.js';
 import {delay} from './tbhelpers.js';
 import TBListener from './tblistener.js';
-import TBLog from './tblog';
 import TBModule from './tbmodule.jsx';
 import {getCache, setCache} from './util/cache';
 import {documentInteractive} from './util/dom';
+import TBLog from './util/logger';
 import {isUserLoggedInQuick} from './util/platform';
 import {getSettingAsync, setSettingAsync, updateSettings} from './util/settings';
 import {reactRenderer} from './util/ui_interop';
