+++ extension/data/modules/achievements.js
@@ -2,8 +2,8 @@ import $ from 'jquery';
 
 import * as TBCore from '../tbcore.js';
 import * as TBHelpers from '../tbhelpers.js';
-import TBLog from '../tblog.ts';
 import {Module} from '../tbmodule.jsx';
+import TBLog from '../util/logger.ts';
 import {getSettingAsync} from '../util/settings.ts';
 
 const log = TBLog('Achievements');
