+++ extension/data/modules/betterbuttons.js
@@ -2,9 +2,9 @@ import $ from 'jquery';
 
 import * as TBApi from '../tbapi.ts';
 import * as TBCore from '../tbcore.js';
-import TBLog from '../tblog.ts';
 import {Module} from '../tbmodule.jsx';
 import {actionButton} from '../tbui.js';
+import TBLog from '../util/logger.ts';
 import {getSettingAsync} from '../util/settings.ts';
 
 const log = TBLog('BButtons');
