+++ extension/data/modules/config.js
@@ -4,11 +4,11 @@ import $ from 'jquery';
 import * as TBApi from '../tbapi.ts';
 import * as TBCore from '../tbcore.js';
 import * as TBHelpers from '../tbhelpers.js';
-import TBLog from '../tblog.ts';
 import {Module} from '../tbmodule.jsx';
 import * as TBui from '../tbui.js';
 import {clearCache} from '../util/cache.ts';
 import {icons} from '../util/icons.ts';
+import TBLog from '../util/logger.ts';
 import {getSettingSync} from '../util/oldLegacyStorageBullshit.ts';
 import {purify, purifyObject} from '../util/purify.js';
