+++ extension/data/modules/domaintagger.js
@@ -3,10 +3,10 @@ import $ from 'jquery';
 import * as TBApi from '../tbapi.ts';
 import * as TBCore from '../tbcore.js';
 import * as TBHelpers from '../tbhelpers.js';
-import TBLog from '../tblog.ts';
 import {Module} from '../tbmodule.jsx';
 import * as TBui from '../tbui.js';
 import {getCache, setCache} from '../util/cache.ts';
+import TBLog from '../util/logger.ts';
 import {purifyObject} from '../util/purify.js';
 
 const log = TBLog('DTagger');
