+++ extension/data/modules/macros.js
@@ -3,9 +3,9 @@ import $ from 'jquery';
 import * as TBApi from '../tbapi.ts';
 import * as TBCore from '../tbcore.js';
 import * as TBHelpers from '../tbhelpers.js';
-import TBLog from '../tblog.ts';
 import {Module} from '../tbmodule.jsx';
 import * as TBui from '../tbui.js';
+import TBLog from '../util/logger.ts';
 import {purify} from '../util/purify.js';
 
 const log = TBLog('ModMacros');
