+++ extension/data/modules/nukecomments.js
@@ -4,9 +4,9 @@ import * as TBApi from '../tbapi.ts';
 import * as TBCore from '../tbcore.js';
 import * as TBHelpers from '../tbhelpers.js';
 import TBListener from '../tblistener.js';
-import TBLog from '../tblog.ts';
 import {Module} from '../tbmodule.jsx';
 import * as TBui from '../tbui.js';
+import TBLog from '../util/logger.ts';
 import {purify, purifyObject} from '../util/purify.js';
 
 const log = TBLog('CommentNuke');
