+++ extension/data/modules/personalnotes.js
@@ -7,7 +7,7 @@ import {Module} from '../tbmodule.jsx';
 import * as TBui from '../tbui.js';
 import {icons} from '../util/icons.ts';
 
-import TBLog from '../tblog.ts';
+import TBLog from '../util/logger.ts';
 import {modbarExists} from './modbar.js';
 
 const log = TBLog('PNotes');
