+++ extension/data/modules/queuetools.js
@@ -4,10 +4,10 @@ import * as TBApi from '../tbapi.ts';
 import * as TBCore from '../tbcore.js';
 import * as TBHelpers from '../tbhelpers.js';
 import TBListener from '../tblistener.js';
-import TBLog from '../tblog.ts';
 import {Module} from '../tbmodule.jsx';
 import * as TBui from '../tbui.js';
 import {getCache, setCache} from '../util/cache.ts';
+import TBLog from '../util/logger.ts';
 import {getSettingSync} from '../util/oldLegacyStorageBullshit.ts';
 import {purifyObject} from '../util/purify.js';
